Operational!
Adding a buffering screen between the button presentations within each comparison of levels helped to alleviate the previous problem (where the 1 vs 1 step was being skipped in comparison 2-6). Data saving is now also changed to include the current value of the buttons and the currect n-back levels that are being compared.