Non-operational: version of nback_COGED_Paradigm with constant variable names
Like for the value iteration, this is a version of the paradigm in which the variable names are the same for every round of n-back, resulting in nicer data files. However, the fixation cross is a blank screen and the first letter is being skipped, which has yet to be fixed.