Final n-back level now chosen from fifth decision step
The final n-back level after the ED part is now chosen from the categories' values at step 5 in the ED-process, not step 6. Plus, both categories of that step are available for being randomly picked. The message displayed on screen before the final n-back is changed, it now displays the full decision that was being picked.